DTFS-4124: Files renamed, Path renamed, heading update and data-cy property name update
