+++ gef-ui/server/routes/index.js
@@ -25,7 +25,7 @@ const facilityConfirmDeletionRoutes = require('./facility-confirm-deletion');
 const returnToMaker = require('./return-to-maker');
 const supportingInformation = require('./supporting-information');
 const file = require('./file');
-const reviewUkefDecision = require('./review-ukef-decision');
+const reviewUkefDecision = require('./review-decision');
 const confirmCoverStartDate = require('./confirm-cover-start-date');
 
 const router = express.Router();
