+++ gef-ui/templates/includes/application-preview/task-comments.njk
@@ -6,7 +6,7 @@
 
     {% if isUkefReviewAvailable %}
       <div id="ukef-review" data-cy="ukef-review">
-        <a href="{{ applicationId }}/review-ukef-decision" class="govuk-link">Review UKEF decision</a>
+        <a href="{{ applicationId }}/review-decision" class="govuk-link">Review UKEF decision</a>
       </div>
     {% endif %}
