+++ gef-ui/templates/index.njk
@@ -29,7 +29,7 @@
 
   <body class="govuk-template__body govuk-body">
 
-    <script src="/assets/js/js-enabled.js" type="text/javascript"></script>
+    <script src="/assets/js/jsEnabled.js" type="text/javascript"></script>
 
     {% include "includes/skip-link.njk" %}
